Adaptando as telas de estoque e produto para o modelo de resposta do servidor

+++ App/app/src/main/java/com/teste/projeto_3/model/Produto.kt
@@ -11,10 +11,9 @@ class Produto(
     var foto: String = "",
     var quantidade: String = "",
     var dataValidade: MutableList<String> = mutableListOf(),
-    var preco_medio: Int = -1,
+    var preco_medio: Double = -1.00,
     var lista_quantidades: MutableList<Int> = mutableListOf(),
 ) : Parcelable {
-
     override fun toString(): String {
         return "Produto(id='$id', nome='$nome', descricao='$descricao', foto='$foto', quantidade='$quantidade', dataValidade='$dataValidade', preco_medio='$preco_medio', lista_quantidades='$lista_quantidades')"
     }
